requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
sqlparse==0.4.4
pytz==2023.3
pytest-django==4.5.2